+++ src/routes/moderation/index.js
@@ -3,13 +3,17 @@ import Requests from './components/Requests'
 import { Loading } from '../../components/loading'
 import useJoinedSpaces from '../../components/matrix_joined_spaces'
 import Matrix from '../../Matrix'
+import { MatrixEvent } from 'matrix-js-sdk'
 import { useTranslation } from 'react-i18next'
 import InviteUserToSpace from './components/InviteUserToSpace'
+import RightsManagement from './components/RightsManagement'
 
 const Moderation = () => {
   const { joinedSpaces, spacesErr, fetchSpaces } = useJoinedSpaces(false)
   const [moderationRooms, setModerationRooms] = useState()
+  const [userSearch, setUserSearch] = useState([])
   const [selection, setSelection] = useState('')
+  const [fetching, setFetching] = useState(false)
   const matrixClient = Matrix.getMatrixClient()
   const { t } = useTranslation()
 
@@ -30,6 +34,28 @@ const Moderation = () => {
     }
   }, [joinedSpaces])
 
+  const setPower = async (roomId, userId, level) => {
+    console.log('changing power level for ' + userId)
+    const stateEvent = await matrixClient.getStateEvent(roomId, 'm.room.power_levels', '')
+    const powerEvent = new MatrixEvent({
+      type: 'm.room.power_levels',
+      content: stateEvent
+    })
+    return matrixClient.setPowerLevel(roomId, userId, level, powerEvent)
+  }
+  const fetchUsers = async (e, search) => {
+    e.preventDefault()
+    setFetching(true)
+    try {
+      const users = await matrixClient.searchUserDirectory({ term: search })
+      // we only update the state if the returned array has entries, to be able to check if users a matrix users or not further down in the code (otherwise the array gets set to [] as soon as you selected an option from the datalist)
+      users.results.length > 0 && setUserSearch(users.results)
+    } catch (err) {
+      console.error('Error whhile trying to fetch users: ' + err)
+    } finally {
+      setFetching(false)
+    }
+  }
   const GetRequestPerRoom = ({ request }) => {
     const room = matrixClient.getRoom(request.room_id)
     // console.log(Object.values(room.currentState.members))
@@ -43,20 +69,16 @@ const Moderation = () => {
     })
   }
 
-  const RightsManagement = () => {
-    return <div>Hello</div>
-  }
   const renderSelection = () => {
     switch (selection) {
       case 'invite':
-        return <InviteUserToSpace matrixClient={matrixClient} moderationRooms={moderationRooms} />
+        return <InviteUserToSpace matrixClient={matrixClient} moderationRooms={moderationRooms} setPower={setPower} fetchUsers={fetchUsers} fetching={fetching} userSearch={userSearch} />
       case 'rights':
-        return <RightsManagement matrixClient={matrixClient} moderationRooms={moderationRooms} />
+        return <RightsManagement moderationRooms={moderationRooms} setPower={setPower} fetchUsers={fetchUsers} fetching={fetching} userSearch={userSearch} />
       default:
         return (
           moderationRooms.length > 0
             ? <>
-
               <section className="requests">
                 {moderationRooms.map((request, index) => <React.Fragment key={request.name}><h3>{request.name}</h3><GetRequestPerRoom request={request} key={index} /></React.Fragment>)}
               </section>
